refactor: split four remaining God-Files into modules
Thin wrappers for market-signals and deals actions; barrel + header/cell modules for reference table columns; section components for reference form content.
