+++ app/dashboard/deals/deal-action-types.ts
@@ -1,0 +1,6 @@
+export type MatchSuggestion = {
+  id: string
+  title: string
+  company_name: string
+  logo_url?: string | null
+}
