+++ app/dashboard/overview/reference-table-column-shared.ts
@@ -1,0 +1,5 @@
+import type * as React from "react"
+
+export function columnWidthStyle(width: number): React.CSSProperties {
+  return { width, minWidth: width, maxWidth: width }
+}
